fix(constants): Adds a per-user cache fallback (#4802)
**Summary**

When Yarn is being run with a different user than the one who created the cache folder, everything
might break (in particular if the cache folder has been created by the root user). This PR fixes this by using the user UID as part of the temporary folder name if possible. If not, it will fallback to the same directory than before.

**Test plan**

The change is small enough that it doesn't need an extra test. We already have a test that makes sure we correctly select a fallback, so adding a new one would only check that it has a specific name, which seems low-signal and prone to break.